Add an option to make the sum of the external potential a constant
This is important when learning the energy from the density.
This is because the density only determines the potential up to
a constant (by the first HK theorem). Therefore, if the sum of
the potential is not constant, then we can end up with a situation
where the same density can come from multiple different potentials,
each of which leads to a different energy. Therefore, there is no
one-to-one map from the density to the energy, and the learning
procedure fails badly. The change introduced in this commit prevents
this, so that the energy can be learnt successfully from the density.