Update ConvNet and training to allow multiple input sizes
ConvNet can now apply a max pooling layer after the final
convolutional layer, so that the output size is independent of
the input size.

The DataLoader has the option to use a collate function which
returns a list of 2d tensors, each for a fixed input size. The
rest of the training function has been modified to handle this
case.